refactor: 💡 Move exceptions for Lecture
