+++ apps/java-api/src/main/java/org/scoalaonline/api/exception/Lecture/LectureInvalidTitleException.java
@@ -1,4 +1,4 @@
-package org.scoalaonline.api.exception;
+package org.scoalaonline.api.exception.Lecture;
 
 /**
  * Exception class for when the Lecture's title is invalid.
